Use SbtRig for Deployments (#79)
Came with a lot of new scalac options that were updated for as well.

Implicit Widening
Procedure Syntax
Classes/Traits in Package Objects
Warn Value Discard

Were the ones that had significant changes as a result. Would like @rossabaker to make sure the release command looks right and that everything is in place to run a deploy.

+++ core/src/main/scala/org/http4s/blaze/pipeline/stages/monitors/ConnectionMonitor.scala
@@ -48,17 +48,17 @@ abstract class ConnectionMonitor {
     }
 
     def writeRequest(data: ByteBuffer): Future[Unit] = {
-      bytesOutBound(data.remaining)
+      bytesOutBound(data.remaining.toLong)
       channelWrite(data)
     }
 
     override def writeRequest(data: Seq[ByteBuffer]): Future[Unit] = {
-      bytesOutBound(data.foldLeft(0)((i, b) => i + b.remaining))
+      bytesOutBound(data.foldLeft(0)((i, b) => i + b.remaining()).toLong)
       channelWrite(data)
     }
 
     def readRequest(size: Int): Future[ByteBuffer] =
-      channelRead(size).map { b => bytesInbound(b.remaining); b}(directec)
+      channelRead(size).map { b => bytesInbound(b.remaining.toLong); b}(directec)
   }
 }
